# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MotoHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      operationId: upsertUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertUser" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUser" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/signin:
    post:
      operationId: signIn
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignInRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /users/{id}/cpf:
    patch:
      operationId: setUserCpf
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetCpfRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: Invalid CPF
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "409":
          description: CPF already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /users/{id}/plan:
    patch:
      operationId: updateUserPlan
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePlan" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    User:
      type: object
      required: [id, name, email, plan, isAdmin, banned, storeName, bio, city, phoneVerified, accountVerified, accountType, acceptedTerms, createdAt]
      properties:
        id: { type: integer }
        publicId: { type: ["string", "null"] }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        cpf: { type: ["string", "null"] }
        cnpj: { type: ["string", "null"] }
        accountType: { type: string, enum: [pessoa, empresa] }
        avatarUrl: { type: ["string", "null"] }
        acceptedTerms: { type: boolean }
        phoneVerified: { type: boolean }
        accountVerified: { type: boolean }
        plan: { type: string, enum: [free, pro, premium] }
        isAdmin: { type: boolean }
        banned: { type: boolean }
        storeName: { type: string }
        bio: { type: string }
        city: { type: string }
        createdAt: { type: string }
    UpsertUser:
      type: object
      required: [name, email, password, phone]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        phone: { type: string, minLength: 10 }
        accountType: { type: ["string", "null"] }
        acceptedTerms: { type: ["boolean", "null"] }
    SignInRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    SetCpfRequest:
      type: object
      required: [cpf]
      properties:
        cpf: { type: string, minLength: 11 }
    UpdateUser:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        storeName: { type: string }
        bio: { type: string }
        city: { type: string }
        banned: { type: boolean }
        accountVerified: { type: boolean }
        isAdmin: { type: boolean }
        plan: { type: string, enum: [free, pro, premium] }
        avatarUrl: { type: string }
        accountType: { type: string }
        cnpj: { type: string }
    UpdatePlan:
      type: object
      required: [plan]
      properties:
        plan: { type: string, enum: [free, pro, premium] }
...