# lib/api-zod/src/generated/types/updatePlan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MotoHub API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePlanPlan } from './updatePlanPlan';

export interface UpdatePlan {
  plan: UpdatePlanPlan;
}