# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MotoHub API
 * OpenAPI spec version: 0.1.0
 */
import type { UserAccountType } from './userAccountType';
import type { UserPlan } from './userPlan';

export interface User {
  id: number;
  /** @nullable */
  publicId?: string | null;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  cpf?: string | null;
  /** @nullable */
  cnpj?: string | null;
  accountType: UserAccountType;
  /** @nullable */
  avatarUrl?: string | null;
  acceptedTerms: boolean;
  phoneVerified: boolean;
  accountVerified: boolean;
  plan: UserPlan;
  isAdmin: boolean;
  banned: boolean;
  storeName: string;
  bio: string;
  city: string;
  createdAt: string;
}